simplified bulkhead = bushing in intake.FCStd
This commit made further progress on the optimization/simplification of
the intake.FCStd file by focusing on its largest objects:

      1,908,593 | Bulkhead Body001
      1,908,590 | Bulkhead Body

Each of these items was about 1.8 MB large. They were unnecessary
complicated parts, complete with threads.

To replace them, I used the Open Source Ecology Piping Workbench:

 * https://wiki.opensourceecology.org/wiki/OSE_Piping_Workbench#Installation

This let me generate a Bushing in one-click (a bushing is another name
for a through-wall bulkhead). The size I selected was identified as part
"837-338" -- which has the following attributes:

 * PartNumber: 837-338
 * PipeSize: 3"
 * PipeSize1: 2"
 * PSize: DN80
 * PSize1: DN50
 * Schedule: 80
 * L: 2+7/32 in
 * Approx Wt: 0.47 lb
 * POD: 3.5 in
 * POD1: 2.375 in
 * PThk1: 0.218 in
 * PID1: 1.913 in

This should be identical to what Jack had added before, at least in
inner and outer diameter. It's what we call a "2 inch pipe"

The savings? Two orders of magnitude:

         41,680 | 2in-bushing-001
         41,680 | 2in-bushing-002

These two parts used-to total to a MemSize of 3.64 MB. Now they total to
0.08 MB. Nice!

TODO: further simplification of any remaning single-objects >1 MB:

     11,045,521 | intake
      1,908,593 | Bulkhead Body001
      1,908,590 | Bulkhead Body
      1,791,548 | 2inPVCValve2
      1,120,897 | Turbulance-Bar_002
      1,120,897 | Turbulance-Bar_001