simplified 2" valves
This commit optimizes the intake.fcstd file by replacing the
overly-complex actuators with simple 2" pvc vaves.

The old MemSize was

-------------------------------
MemSize (bytes) | Object Label
-------------------------------
      6,603,525 | intake
      1,791,548 | 2inPVCValve2
      1,120,897 | Turbulance-Bar_002
      1,120,897 | Turbulance-Bar_001

And the new MemSize is

-------------------------------
MemSize (bytes) | Object Label
-------------------------------
      3,325,258 | intake
      1,120,897 | Turbulance-Bar_002
      1,120,897 | Turbulance-Bar_001
        180,035 | intake_Outflow_Pipe_001
        180,035 | intake_Outflow_Pipe_002
        155,122 | 2in-pvc-valve-001
        155,122 | 2in-pvc-valve-002

Therefore, we were able to reduce the size from 1.71 MB to 0.15 MB, a
savings of 1 magnitude

Heuristically, I can now say that the FreeCAD finally opens this file
*much* faster -- with significantly less lag/waiting after opening
before the model appears.

TODO: replace the outflow pipes with threadless, simplified objects. And
also the turbulance bars should be simplified (it's unclear why these
simple angle bars are over 1MB each -- perhaps it's the threads on the
bolts)