simplified turblence bars
This commit deleted the overly-complex model of the turbulance bar (and
its two bolts) to reduce the size of the inktake.fcstd file and optimize
future edits.

I measured the previously-existing steel angle to be 650mm long, about
37mm wide (from the 90 degree angle), and about 5.75 mm thick.

I did a quick check for common angle sizes in metric, and I found 60 mm
x 6 mm to be a common size. I decided to use this. No further research
has been done to decide on this size. It could be sufficient. Maybe
something thinner would work too. Actual testing would be needed. We're
just in the design phase.

Anyway, I just did a simple sketch for the angle and extruded a 3D
object for it.

For the bolts, I generated a M16 hex bolt that's 40mm long using the
BOLTS addon for FreeCAD

 * https://github.com/boltsparts/BOLTSFC/blob/main/README.md

The official FreeCAD project on GitHub also had some bolts, but they
didn't have a hex head that we needed. In any case, the docs of this
repo made it clear that we should *not* be modeling fasteners with
threads, as it just causes file (and repo) bloat and slows down FreeCAD
unnecessarily (harming future dev).

> In most CAD systems, screw threads are never modelled because rendering many threaded screws in the same document would be too CPU intensive and would serve no purpose. These screws are intended to be used in assemblies and as such they are represented as cylindrical, with the nominal screw size as their outer diameter.

 * https://github.com/FreeCAD/FreeCAD-library/tree/master/Mechanical%20Parts/Fasteners/

Here was the MemSize before these changes

-------------------------------
MemSize (bytes) | Object Label
-------------------n------------
      3,325,258 | intake
      1,120,897 | Turbulance-Bar_002
      1,120,897 | Turbulance-Bar_001
        180,035 | intake_Outflow_Pipe_001
        180,035 | intake_Outflow_Pipe_002
        155,122 | 2in-pvc-valve-001
        155,122 | 2in-pvc-valve-002

And now after:

MemSize (bytes) | Object Label
-------------------------------
      1,166,286 | intake
        180,035 | intake_Outflow_Pipe_001
        180,035 | intake_Outflow_Pipe_002
        155,122 | 2in-pvc-valve-001
        155,122 | 2in-pvc-valve-002
         52,730 | Hole001
         48,855 | Thickness001
         41,680 | 2in-bushing-001
         41,680 | 2in-bushing-002
         37,591 | Hatch
         31,366 | hex-bolt-M16-40mm-001
         31,366 | hex-bolt-M16-40mm-002
         28,969 | weir (notch top)
         21,847 | Hole
         21,534 | Thickness
         21,517 | intake drum
         20,269 | turbulance-bar-001

Therefore, the original turbulance bar was 1.07 MB. The new size is
0.08 MB. Win!

Oh, there was also a second "turbulence bar" that was copied on the
drum-side of the weir. I'm 80% sure this was added to catch water that
falls in-between the weir and the drum, to funnel it onto the screen and
into the drum. I deleted this.

At the time of writing, We've had two prototypes of this intake system
built in Ecuador. In
one, a ton of water is lost due to this gap, which is especially a
problem in dry season--when there's just a trickle of water coming over
the weir. In the other, they put chicken wire on the top of the weir and
cemented over it. That effectively created a "ledge" that forces the
water to fall further away from the weir -- directly over the mesh ontop
of the drum. I think this is a better design. In the future, we should
update the weir top to have a "ledge" that sticks-out. I'd also like to
change the weir opening to a V-shape, so then we can "close the gap"
with a very small "ledge" at the bottom of the V that will capture water
in low-flow as-needed.