updated images
This commit updates the main image of the intake
(life-line_2025.10.jpg), since I've made the numerous optimizations to
the file in the last week.

I also made an SVG file that takes the above raster image and adds some
numbers and arrows to identify different parts. I can use this in the
documentation to describe, in a few sentences, what each of these items
are.

To the end above, I also captured an image at a different angle for each
of the 7 numbered items in the svg file. I further edited these photos
in GIMP for clairy, to highlight which component is which

This time I got better (higher definition) images from FreeCAD via Tools -> Save Picture by clicking the "Extended" button and then changing the "Standard sizes" from "Current screen" to "QUXHA" (3,200 x 2,400 px)

TODO: actually update the documentation to use these images