lots of docs changes
This commit makes a lot of changes to the sphinx documentation, most
importantly to the faq section and the intake "build instructions"
section.

We now have an svg image with 7 numbers pointing to 7 different
sub-components of the "intake". And at least 2 paragraphs and a photo
for each of these

TODO: Bill of Materials and real-world example photos (next to each of
the 3D model screenshots)