fixed docs bom for intake
This commit adds a new file 'bom.ods'. It makes more sense long-term to
use a proper spreadsheet for a BOM, and this one uses the open-source
LibreOffice format. The format isn't great for git, but with tools like
ReZipDoc, we can get git to unzip FreeCAD and LibreOffice files, to
save repo space and get better diffs

 * https://github.com/hoijui/ReZipDoc

For now, I just exported the contents of the .ods file to a .csv file,
and used a quick online converter tool to have it convert the csv to
reST format

 * https://tableconvert.com/csv-to-restructuredtext

In the future, we should write a sphinx directive that can import the
.ods file directly into the documentation at build-time

 * https://softwarerecs.stackexchange.com/questions/75204/documentation-as-code-markup-supporting-tables-with-calculations-basic-spreads

...but for now, I think this is the best solution