documentation improvements
This commit did some cleanup of the main "what is the life-line?"
section, fixing the disordered heading syntax and setting :titlesonly:
in the toc (preventing subheadings from appearing in the left-hand
navbar).

I also updated the main repo's README with a link to the docs, and
deleted the now-redundant info.