trivial change
This commit makes just a trivial change to the CHANGELOG (adding a URL
to the article soon-to-be-published announcing the release).

What I'm attempting to do is to push a rebuild of the docs from the
'main' branch, because (for some reason) the 'v2025.10' git tag that I
created this morning did *not* add a 'v2025.10' "version" in the
left-hand navbar of the sphinx documentation.

I'm hoping that a push on 'main' (instead of the branch) will fix that.