resize logo to 150px
This commit adds a '_static/custom.css' file that (for now) just
expliclty sets the logo in the left-hand navbar to 150px wide.

In a recent commit, we switched from svg to png, to ensure the font is
consistant in all user agents. A side-effect of that is that the image
grew to a silly size. This is much more reasonable size.