lowered the weir to match the height of the intake barrel (plus 20 cm).
Our docs indicate that the barrel needs to be located at least 20 cm
below the weir opening, but we had some math in-place for the height of
the weir (notch bottom) to be 585 cm high plus 20 cm. But the height of
the top-edge of the downsloped intake barrel wasn't at a base of 585 cm
off the ground. It's only about 536 mm off the ground. So I fixed that.