created spreadsheet
For a more robust design, I'm building on the last commit by putting
some of these dimensions into a spreadsheet. And the following objects
now reference the spreadsheet, to determine their size and/or placement:

1. Weir (notch bottom)
2. Weir (notch top)
3. Turbulance Bar
4. Bolts (for turbulance bar)

So now we can just update the spreadsheet, and those other objects
should dynamically all update, as desired.

TODO: cut a V-shape out of the weir opening && add a "lip"