extruded triangle cutout
This commit finished extruding and raising-up the triangle cutout of our
weir, but I made a mistake: the triangle is upside down! We want the
point of the triangle to be facing down!

TODO: re-draw the triangle, so that the point is downward, directing the
smallest flow of water along a single line, then increasing its spread
as the flow increases