cut the triangle out of the weir
This commit finishes the "cut" that subtracts the upside-down triangle
out of the weir wall, to creat a triangular weir opening. This change
increases the efficiency of the water collection in low-rain/low-flow conditions, such that a slow trickle is directed at a single point over the drum.

TODO: add a "lip" to ensure that 100% of the water coming out of the
weir opening in low-flow conditions gets dropped directly over the drum